app completa

+++ views/admin/instructores/lista_instructores.php
@@ -4,6 +4,10 @@ $db = new Database();
 $conectar = $db->conectar();
 session_start();
 
+if (!isset($_SESSION['documento'])) {
+    header("Location: ../../../login.php"); // Redirigir a la página de inicio si no está logueado
+    exit();
+}
 // Verifica si la clave 'documento' está definida en la sesión antes de usarla
 if (isset($_SESSION['documento'])) {
     $documento = $_SESSION['documento'];
@@ -75,35 +79,63 @@ if (isset($_SESSION['documento'])) {
                         <div class="full">
                             <div class="center-desk">
                                 <div class="logo">
-                                    <a href="index.html"><img src="../../../images/Sena_Colombia_logo.svg.png" alt="#" /></a>
+                                    <a href="#"><img src="../../../images/Sena_Colombia_logo.svg.png" alt="#" /></a>
                                 </div>
-                                <h2 class="titulo-principal" style="color:#000;">Bienvenido Administrador <?= $usuario['nombre']; ?> </h2>
                             </div>
                         </div>
+                    </div>
+                    <div class="col-md-10 offset-md-1">
+                        <nav class="navigation navbar navbar-expand-md navbar-dark ">
+                            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExample04" aria-controls="navbarsExample04" aria-expanded="false" aria-label="Toggle navigation">
+                                <span class="navbar-toggler-icon"></span>
+                            </button>
+                            <div class="collapse navbar-collapse" id="navbarsExample04">
+                                <ul class="navbar-nav mr-auto">
+
+                                    <li class="nav-item active">
+                                        <a class="nav-link" href="#">Lista de instructores </a>
+                                    </li>
+                                    <li class="nav-item">
+                                        <a class="nav-link" href="lista_instructores_asociados.php">Lista de instructores asociados</a>
+                                    </li>
+                                </ul>
+                            </div>
+                        </nav>
                     </div>
                 </div>
             </div>
         </div>
     </header>
+    <style>
+        .search-input {
+            border: 1px solid #6c757d;
+            /* Color de borde más oscuro */
+            border-radius: 0.25rem;
+            /* Bordes redondeados */
+            color: #495057;
+            /* Color de texto */
+        }
+    </style>
+
 
     <div class="container mt-3">
         <a href="crear_instructor.php" class="btn btn-success mb-2">Crea un instructor</a>
+        <a href="asignacion_instructor.php" class="btn btn-primary mb-2">Asignar instructor</a>
 
-        <a href="asignacion_instructor.php" class="btn btn-primary mb-2">asignar instructor</a>
+        <input type="text" id="searchInput" class="form-control mb-2 search-input" placeholder="Buscar por nombre o documento">
 
-        <table class="table table-striped table-bordered table-hover">
+        <table id="instructorTable" class="table table-striped table-bordered table-hover">
             <thead class="thead-dark">
                 <tr style="text-transform: uppercase;">
                     <th>Documento</th>
                     <th>Nombre</th>
-                    <th>Direccion</th>
+                    <th>Dirección</th>
                     <th>Empresa</th>
                     <th>Estado</th>
                     <th>Acciones</th>
                 </tr>
             </thead>
             <tbody>
-
                 <?php foreach ($asigna as $usua) { ?>
                     <tr>
                         <td><?= $usua["documento"] ?></td>
@@ -113,55 +145,51 @@ if (isset($_SESSION['documento'])) {
                         <td><?= $usua["estado"] ?></td>
                         <td>
                             <a href="editar_instructor.php?id=<?= $usua["documento"] ?>" class="btn btn-primary ">Actualizar</a>
-                            <a href="eliminar_instructor.php?id=<?= $usua["documento"] ?>" class="btn btn-danger">Eliminar</a>
+                            <a href="eliminar_instructor.php?id=<?= $usua["documento"] ?>" class="btn btn-danger" onclick="return confirm('¿Estás seguro de que quieres eliminar este instructor?')">Eliminar</a>
                         </td>
                     </tr>
                 <?php } ?>
             </tbody>
         </table>
+
         <a href="../index.php" class="btn btn-danger">Regresar</a>
+        <a href="exportar_pdf.php" class="btn btn-primary">Generar Reporte</a>
     </div>
+
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+    <script src="../../../paginacion_buscados.js"> </script>
     <!-- footer -->
     <footer>
         <div class="footer">
             <div class="container">
                 <div class="row">
                     <div class=" col-md-3 col-sm-6">
-                        <ul class="social_icon">
-                            <li><a href="#"><i class="fa fa-instagram" aria-hidden="true"></i></a></li>
-                        </ul>
-                        <p class="variat pad_roght2">There are many variat
-                            ions of passages of L
-                            orem Ipsum available
-                            , but the majority h
-                            ave suffered altera
-                            tion in some form, by
+                        <h3>variedad</h3>
+                        <p class="variat pad_roght2">Ofrecemos una amplia variedad de herramientas
+                            de alta calidad para satisfacer todas tus necesidades de
+                            construcción.Tenemos todo lo que necesitas para completar
+                            tus proyectos con éxito.
                         </p>
                     </div>
                     <div class=" col-md-3 col-sm-6">
-                        <h3>LET US HELP YOU </h3>
-                        <p class="variat pad_roght2">There are many variat
-                            ions of passages of L
-                            orem Ipsum available
-                            , but the majority h
-                            ave suffered altera
-                            tion in some form, by
+                        <h3>dejanos ayudarte </h3>
+                        <p class="variat pad_roght2">Nuestro objetivo es facilitarte el acceso a las herramientas
+                            que necesitas para tus proyectos. Con nuestro proceso de préstamo simple y transparente,
+                            puedes obtener las herramientas adecuadas sin complicaciones ni demoras.
                         </p>
                     </div>
                     <div class="col-md-3 col-sm-6">
-                        <h3>INFORMATION</h3>
-                        <ul class="link_menu">
-                        </ul>
-                    </div>
-                    <div class="col-md-3 col-sm-6">
-                        <h3>OUR Design</h3>
-                        <p class="variat">There are many variat
-                            ions of passages of L
-                            orem Ipsum available
-                            , but the majority h
-                            ave suffered altera
-                            tion in some form, by
+                        <h3>NUESTRO DISEÑO</h3>
+                        <p class="variat">En nuestra empresa, nos esforzamos por ofrecer un diseño intuitivo
+                            y fácil de usar en todas nuestras plataformas. Nuestra interfaz está diseñada
+                            pensando en la comodidad y la accesibilidad del usuario.
                         </p>
+                    </div>
+                    <div class="col-md-6 offset-md-6">
+                        <form id="hkh" class="bottom_form">
+                            <input class="enter" placeholder="" type="text" name="Enter your email">
+                            <button class="sub_btn">Prestamos de herramientas</button>
+                        </form>
                     </div>
                 </div>
             </div>
@@ -169,7 +197,7 @@ if (isset($_SESSION['documento'])) {
                 <div class="container">
                     <div class="row">
                         <div class="col-md-10 offset-md-1">
-                            <p>© 2019 All Rights Reserved. Design by <a href="https://html.design/"> Free Html Templates</a></p>
+                            <p>© 2019 All Rights Reserved. Design by <a href="https://html.design/"> Cristian Figueroa</a></p>
                         </div>
                     </div>
                 </div>
